split staging components out of agent
