+++ src/radical/pilot/agent/__init__.py
@@ -3,8 +3,10 @@ __copyright__ = "Copyright 2013-2014, http://radical.rutgers.edu"
 __license__   = "MIT"
 
 
-from .rm        import RM
-from .lm        import LaunchMethod as LM
-from .scheduler import Scheduler
-from .executing import Executing
+from .rm             import RM
+from .lm             import LaunchMethod as LM
+from .scheduler      import Scheduler
+from .executing      import Executing
+from .staging_input  import Input
+from .staging_output import Output
