+++ src/radical/pilot/agent/radical-pilot-agent-multicore.py
@@ -364,324 +364,6 @@ def pilot_DONE(mongo_p=None, pilot_uid=None, logger=None, msg=None):
 #
 # ==============================================================================
 #
-class AgentStagingInputComponent(rpu.Component):
-    """
-    This component performs all agent side input staging directives for compute
-    units.  It gets units from the agent_staging_input_queue, in
-    AGENT_STAGING_INPUT_PENDING state, will advance them to AGENT_STAGING_INPUT
-    state while performing the staging, and then moves then to the
-    AGENT_SCHEDULING_PENDING state, into the agent_scheduling_queue.
-    """
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg):
-
-        rpu.Component.__init__(self, 'AgentStagingInputComponent', cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def create(cls, cfg):
-
-        return cls(cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def initialize_child(self):
-
-        self.declare_input (rp.AGENT_STAGING_INPUT_PENDING, rp.AGENT_STAGING_INPUT_QUEUE)
-        self.declare_worker(rp.AGENT_STAGING_INPUT_PENDING, self.work)
-
-        self.declare_output(rp.ALLOCATING_PENDING, rp.AGENT_SCHEDULING_QUEUE)
-
-        self.declare_publisher('state', rp.AGENT_STATE_PUBSUB)
-
-        # all components use the command channel for control messages
-        self.declare_publisher ('command', rp.AGENT_COMMAND_PUBSUB)
-
-        # communicate successful startup
-        self.publish('command', {'cmd' : 'alive',
-                                 'arg' : self.cname})
-
-
-    # --------------------------------------------------------------------------
-    #
-    def finalize_child(self):
-
-        # communicate finalization
-        self.publish('command', {'cmd' : 'final',
-                                 'arg' : self.cname})
-
-
-
-    # --------------------------------------------------------------------------
-    #
-    def work(self, cu):
-
-        self.advance(cu, rp.AGENT_STAGING_INPUT, publish=True, push=False)
-        self._log.info('handle %s' % cu['_id'])
-
-        workdir      = os.path.join(self._cfg['workdir'], '%s' % cu['_id'])
-        gtod         = os.path.join(self._cfg['workdir'], 'gtod')
-        staging_area = os.path.join(self._cfg['workdir'], self._cfg['staging_area'])
-        staging_ok   = True
-
-        cu['workdir']     = workdir
-        cu['stdout']      = ''
-        cu['stderr']      = ''
-        cu['opaque_clot'] = None
-        # TODO: See if there is a more central place to put this
-        cu['gtod']        = gtod
-
-        stdout_file       = cu['description'].get('stdout')
-        stdout_file       = stdout_file if stdout_file else 'STDOUT'
-        stderr_file       = cu['description'].get('stderr')
-        stderr_file       = stderr_file if stderr_file else 'STDERR'
-
-        cu['stdout_file'] = os.path.join(workdir, stdout_file)
-        cu['stderr_file'] = os.path.join(workdir, stderr_file)
-
-        # create unit workdir
-        rpu.rec_makedir(workdir)
-        self._prof.prof('unit mkdir', uid=cu['_id'])
-
-        try:
-            for directive in cu['Agent_Input_Directives']:
-
-                self._prof.prof('Agent input_staging queue', uid=cu['_id'],
-                         msg="%s -> %s" % (str(directive['source']), str(directive['target'])))
-
-                # Perform input staging
-                self._log.info("unit input staging directives %s for cu: %s to %s",
-                               directive, cu['_id'], workdir)
-
-                # Convert the source_url into a SAGA Url object
-                source_url = rs.Url(directive['source'])
-
-                # Handle special 'staging' scheme
-                if source_url.scheme == self._cfg['staging_scheme']:
-                    self._log.info('Operating from staging')
-                    # Remove the leading slash to get a relative path from the staging area
-                    rel2staging = source_url.path.split('/',1)[1]
-                    source = os.path.join(staging_area, rel2staging)
-                else:
-                    self._log.info('Operating from absolute path')
-                    source = source_url.path
-
-                # Get the target from the directive and convert it to the location
-                # in the workdir
-                target = directive['target']
-                abs_target = os.path.join(workdir, target)
-
-                # Create output directory in case it doesn't exist yet
-                rpu.rec_makedir(os.path.dirname(abs_target))
-
-                self._log.info("Going to '%s' %s to %s", directive['action'], source, abs_target)
-
-                if   directive['action'] == LINK: os.symlink     (source, abs_target)
-                elif directive['action'] == COPY: shutil.copyfile(source, abs_target)
-                elif directive['action'] == MOVE: shutil.move    (source, abs_target)
-                else:
-                    # FIXME: implement TRANSFER mode
-                    raise NotImplementedError('Action %s not supported' % directive['action'])
-
-                log_message = "%s'ed %s to %s - success" % (directive['action'], source, abs_target)
-                self._log.info(log_message)
-
-        except Exception as e:
-            self._log.exception("staging input failed -> unit failed")
-            staging_ok = False
-
-
-        # Agent input staging is done (or failed)
-        if staging_ok:
-          # self.advance(cu, rp.AGENT_SCHEDULING_PENDING, publish=True, push=True)
-            self.advance(cu, rp.ALLOCATING_PENDING, publish=True, push=True)
-        else:
-            self.advance(cu, rp.FAILED, publish=True, push=False)
-
-
-# ==============================================================================
-#
-class AgentStagingOutputComponent(rpu.Component):
-    """
-    This component performs all agent side output staging directives for compute
-    units.  It gets units from the agent_staging_output_queue, in
-    AGENT_STAGING_OUTPUT_PENDING state, will advance them to
-    AGENT_STAGING_OUTPUT state while performing the staging, and then moves then
-    to the UMGR_STAGING_OUTPUT_PENDING state, which at the moment requires the
-    state change to be published to MongoDB (no push into a queue).
-
-    Note that this component also collects stdout/stderr of the units (which
-    can also be considered staging, really).
-    """
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg):
-
-        rpu.Component.__init__(self, 'AgentStagingOutputComponent', cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def create(cls, cfg):
-
-        return cls(cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def initialize_child(self):
-
-        self.declare_input (rp.AGENT_STAGING_OUTPUT_PENDING, rp.AGENT_STAGING_OUTPUT_QUEUE)
-        self.declare_worker(rp.AGENT_STAGING_OUTPUT_PENDING, self.work)
-
-        # we don't need an output queue -- units are picked up via mongodb
-        self.declare_output(rp.PENDING_OUTPUT_STAGING, None) # drop units
-
-        self.declare_publisher('state', rp.AGENT_STATE_PUBSUB)
-
-        # all components use the command channel for control messages
-        self.declare_publisher ('command', rp.AGENT_COMMAND_PUBSUB)
-
-        # communicate successful startup
-        self.publish('command', {'cmd' : 'alive',
-                                 'arg' : self.cname})
-
-
-    # --------------------------------------------------------------------------
-    #
-    def finalize_child(self):
-
-        # communicate finalization
-        self.publish('command', {'cmd' : 'final',
-                                 'arg' : self.cname})
-
-
-    # --------------------------------------------------------------------------
-    #
-    def work(self, cu):
-
-        self.advance(cu, rp.AGENT_STAGING_OUTPUT, publish=True, push=False)
-
-        staging_area = os.path.join(self._cfg['workdir'], self._cfg['staging_area'])
-        staging_ok   = True
-
-        workdir = cu['workdir']
-
-        ## parked from unit state checker: unit postprocessing
-        if os.path.isfile(cu['stdout_file']):
-            with open(cu['stdout_file'], 'r') as stdout_f:
-                try:
-                    txt = unicode(stdout_f.read(), "utf-8")
-                except UnicodeDecodeError:
-                    txt = "unit stdout contains binary data -- use file staging directives"
-
-                cu['stdout'] += rpu.tail(txt)
-
-        if os.path.isfile(cu['stderr_file']):
-            with open(cu['stderr_file'], 'r') as stderr_f:
-                try:
-                    txt = unicode(stderr_f.read(), "utf-8")
-                except UnicodeDecodeError:
-                    txt = "unit stderr contains binary data -- use file staging directives"
-
-                cu['stderr'] += rpu.tail(txt)
-
-        if 'RADICAL_PILOT_PROFILE' in os.environ:
-            if os.path.isfile("%s/PROF" % cu['workdir']):
-                try:
-                    with open("%s/PROF" % cu['workdir'], 'r') as prof_f:
-                        txt = prof_f.read()
-                        for line in txt.split("\n"):
-                            if line:
-                                x1, x2, x3 = line.split()
-                                self._prof.prof(x1, msg=x2, timestamp=float(x3), uid=cu['_id'])
-                except Exception as e:
-                    self._log.error("Pre/Post profiling file read failed: `%s`" % e)
-
-        # NOTE: all units get here after execution, even those which did not
-        #       finish successfully.  We do that so that we can make
-        #       stdout/stderr available for failed units.  But at this point we
-        #       don't need to advance those units anymore, but can make them
-        #       final.
-        if cu['target_state'] != rp.DONE:
-            self.advance(cu, cu['target_state'], publish=True, push=False)
-            return
-
-
-        try:
-            # all other units get their (expectedly valid) output files staged
-            for directive in cu['Agent_Output_Directives']:
-
-                self._prof.prof('Agent output_staging', uid=cu['_id'],
-                         msg="%s -> %s" % (str(directive['source']), str(directive['target'])))
-
-                # Perform output staging
-                self._log.info("unit output staging directives %s for cu: %s to %s",
-                        directive, cu['_id'], workdir)
-
-                # Convert the target_url into a SAGA Url object
-                target_url = rs.Url(directive['target'])
-
-                # Handle special 'staging' scheme
-                if target_url.scheme == self._cfg['staging_scheme']:
-                    self._log.info('Operating from staging')
-                    # Remove the leading slash to get a relative path from
-                    # the staging area
-                    rel2staging = target_url.path.split('/',1)[1]
-                    target = os.path.join(staging_area, rel2staging)
-                else:
-                    self._log.info('Operating from absolute path')
-                    # FIXME: will this work for TRANSFER mode?
-                    target = target_url.path
-
-                # Get the source from the directive and convert it to the location
-                # in the workdir
-                source = str(directive['source'])
-                abs_source = os.path.join(workdir, source)
-
-                # Create output directory in case it doesn't exist yet
-                # FIXME: will this work for TRANSFER mode?
-                rpu.rec_makedir(os.path.dirname(target))
-
-                self._log.info("Going to '%s' %s to %s", directive['action'], abs_source, target)
-
-                if directive['action'] == LINK:
-                    # This is probably not a brilliant idea, so at least give a warning
-                    os.symlink(abs_source, target)
-                elif directive['action'] == COPY:
-                    shutil.copyfile(abs_source, target)
-                elif directive['action'] == MOVE:
-                    shutil.move(abs_source, target)
-                else:
-                    # FIXME: implement TRANSFER mode
-                    raise NotImplementedError('Action %s not supported' % directive['action'])
-
-                log_message = "%s'ed %s to %s - success" %(directive['action'], abs_source, target)
-                self._log.info(log_message)
-
-        except Exception as e:
-            self._log.exception("staging output failed -> unit failed")
-            staging_ok = False
-
-
-        # Agent output staging is done (or failed)
-        if staging_ok:
-          # self.advance(cu, rp.UMGR_STAGING_OUTPUT_PENDING, publish=True, push=True)
-            self.advance(cu, rp.PENDING_OUTPUT_STAGING, publish=True, push=False)
-        else:
-            self.advance(cu, rp.FAILED, publish=True, push=False)
-
-
-
-# ==============================================================================
-#
 class AgentWorker(rpu.Worker):
 
     # --------------------------------------------------------------------------
@@ -1120,10 +802,10 @@ class AgentWorker(rpu.Worker):
         # We use a static map from component names to class types for now --
         # a factory might be more appropriate (FIXME)
         cmap = {
-            "AgentStagingInputComponent"  : AgentStagingInputComponent,
+            "AgentStagingInputComponent"  : rp.agent.Input,
             "AgentSchedulingComponent"    : rp.agent.Scheduler,
             "AgentExecutingComponent"     : rp.agent.Executing,
-            "AgentStagingOutputComponent" : AgentStagingOutputComponent
+            "AgentStagingOutputComponent" : rp.agent.Output,
             }
         for cname, cnum in self._sub_cfg.get('components',{}).iteritems():
             for i in range(cnum):
@@ -1582,3 +1264,4 @@ if __name__ == "__main__":
 
 #
 # ------------------------------------------------------------------------------
+
