+++ src/radical/pilot/agent/staging_output/__init__.py
@@ -1,0 +1,8 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import AgentStagingOutputComponent as Output
+
+
